Move file-list state to separate package
