+++ web/src/application/state/initialState.js
@@ -1,0 +1,13 @@
+import collInitialState from "../../collection/state/initialState";
+
+/**
+ * The entire application initial state.
+ */
+const initialState = {
+  /**
+   * File collection management state.
+   */
+  coll: collInitialState,
+};
+
+export default initialState;
