+++ web/src/collection/components/FileBrowserPage/CategorySelector/index.js
@@ -1,2 +1,2 @@
 export { default } from "./CategorySelector";
-export { MatchCategory } from "../../../state/MatchCategory";
+export { MatchCategory } from "../../../state/fileList/MatchCategory";
