+++ web/src/collection/components/FileBrowserPage/FileBrowserActions/ViewSelector.js
@@ -4,7 +4,7 @@ import ListIcon from "@material-ui/icons/ViewStream";
 import GridIcon from "@material-ui/icons/ViewModule";
 import { useIntl } from "react-intl";
 import IconSelect from "../../../../common/components/IconSelect";
-import FileListType from "../../../state/FileListType";
+import FileListType from "../../../state/fileList/FileListType";
 
 function useMessages() {
   const intl = useIntl();
