+++ web/src/collection/components/FileBrowserPage/FileBrowserPage.js
@@ -11,13 +11,13 @@ import FileLinearList from "./FileLinearList/FileLinearList";
 import FileGridList from "./FileGridList";
 import { useDispatch, useSelector } from "react-redux";
 import {
-  selectCounts,
-  selectError,
+  selectFileList,
+  selectFileCounts,
+  selectFileError,
   selectFiles,
-  selectFilters,
-  selectLoading,
+  selectFileFilters,
+  selectFileLoading,
 } from "../../state/selectors";
-import { fetchFiles, selectColl, updateFilters } from "../../state";
 import Fab from "@material-ui/core/Fab";
 import Zoom from "@material-ui/core/Zoom";
 import VisibilitySensor from "react-visibility-sensor";
@@ -25,9 +25,13 @@ import { scrollIntoView } from "../../../common/helpers/scroll";
 import { useHistory, useLocation } from "react-router-dom";
 import { routes } from "../../../routing/routes";
 import { useIntl } from "react-intl";
-import { defaultFilters } from "../../state/reducers";
-import FileListType from "../../state/FileListType";
-import { changeFileListView } from "../../state/actions";
+import FileListType from "../../state/fileList/FileListType";
+import {
+  changeFileListView,
+  fetchFiles,
+  updateFilters,
+} from "../../state/fileList/actions";
+import { defaultFilters } from "../../state/fileList/initialState";
 
 const useStyles = makeStyles((theme) => ({
   container: {
@@ -127,17 +131,17 @@ function FileBrowserPage(props) {
   const { className } = props;
   const classes = useStyles();
   const [showFilters, setShowFilters] = useState(false);
-  const collState = useSelector(selectColl);
-  const error = useSelector(selectError);
-  const loading = useSelector(selectLoading);
+  const fileListState = useSelector(selectFileList);
+  const error = useSelector(selectFileError);
+  const loading = useSelector(selectFileLoading);
   const files = useSelector(selectFiles);
-  const filters = useSelector(selectFilters);
-  const counts = useSelector(selectCounts);
+  const filters = useSelector(selectFileFilters);
+  const counts = useSelector(selectFileCounts);
   const dispatch = useDispatch();
   const [top, setTop] = useState(true);
   const topRef = useRef(null);
   const history = useHistory();
-  const view = collState.fileListType;
+  const view = fileListState.fileListType;
   const List = listComponent(view);
   const intl = useIntl();
   const showFiltersRef = useRef();
@@ -146,10 +150,10 @@ function FileBrowserPage(props) {
   const activeFilters = FilterPane.useActiveFilters();
 
   useEffect(() => {
-    if (!keepFilters || collState.neverLoaded) {
+    if (!keepFilters || fileListState.neverLoaded) {
       dispatch(updateFilters(defaultFilters));
     }
-  }, [keepFilters, collState.neverLoaded]);
+  }, [keepFilters, fileListState.neverLoaded]);
 
   const handleFetchPage = useCallback(() => dispatch(fetchFiles()), []);
