--- web/src/collection/state/index.js
@@ -1,13 +1,0 @@
-export {
-  ACTION_FETCH_FILES_FAILURE,
-  ACTION_FETCH_FILES,
-  ACTION_FETCH_FILES_SUCCESS,
-  ACTION_UPDATE_FILTERS,
-  fetchFiles,
-  fetchFilesFailure,
-  fetchFilesSuccess,
-  updateFilters,
-} from "./actions";
-export { initialState, collRootReducer } from "./reducers";
-export { default as collRootSaga } from "./sagas";
-export { selectColl } from "./selectors";
